@@ -2,44 +2,52 @@
   "name": "mcad-email-tasks",
   "version": "2.0.1",
   "description": "MCAD Email Tasks will generate and send emails",
+  "type": "module",
   "bin": {
     "mcad-email-tasks": "./cli.js"
   },
   "scripts": {
-    "test:watch": "jest --watch",
-    "test": "jest",
+    "test": "node --experimental-vm-modules node_modules/.bin/jest",
+    "test:watch": "node --experimental-vm-modules node_modules/.bin/jest --watch",
     "dev": "nodemon ./cli.js --watch . --ext 'js,json,hbs,css,sql'",
     "start": "./cli.js"
   },
-  "author": "",
+  "engines": {
+    "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+  },
+  "author": "James Johnson <johnsojr@gmail.com>",
   "license": "ISC",
   "devDependencies": {
-    "eslint": "7.22.0",
+    "@jest/globals": "^27.0.3",
+    "eslint": "7.27.0",
     "eslint-config-airbnb-base": "14.2.1",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-prettier": "3.3.1",
-    "husky": "5.2.0",
-    "jest": "26.6.3",
-    "prettier": "2.2.1"
+    "eslint-config-prettier": "8.3.0",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-prettier": "3.4.0",
+    "husky": "6.0.0",
+    "jest": "27.0.4",
+    "prettier": "2.3.0"
   },
   "dependencies": {
     "axios": "^0.21.1",
-    "dotenv": "^8.2.0",
-    "email-templates": "8.0.4",
+    "dotenv": "^10.0.0",
+    "email-templates": "8.0.7",
     "handlebars": "4.7.7",
-    "inquirer": "8.0.0",
-    "lit-html": "1.3.0",
-    "luxon": "1.26.0",
-    "meow": "^9.0.0",
-    "mssql": "6.3.1",
-    "node-sass": "5.0.0",
+    "inquirer": "8.1.0",
+    "lit-html": "1.4.1",
+    "luxon": "1.27.0",
+    "meow": "^10.0.0",
+    "mssql": "7.1.0",
+    "node-sass": "6.0.0",
     "nodemon": "2.0.7",
     "ow": "0.23.0",
-    "postmark": "2.7.3",
+    "postmark": "2.7.4",
     "preview-email": "3.0.4",
     "ramda": "0.27.1",
     "react": "17.0.2",
     "write": "^2.0.0"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
