@@ -1,8 +1,8 @@
-const R = require('ramda');
-const getSundayBefore = require('./getSundayBefore');
-const toPrettyDate = require('./toPrettyDate');
-const getCourseId = require('./getCourseId');
-const getFullTermName = require('./getFullTermName');
+import { pipe } from 'ramda';
+import getSundayBefore from './getSundayBefore.js';
+import toPrettyDate from './toPrettyDate.js';
+import getCourseId from './getCourseId.js';
+import getFullTermName from './getFullTermName.js';
 
 const keyExists = (key, obj) => key in obj;
 
@@ -23,30 +23,34 @@ const includeCourseId = (obj) => {
   return obj;
 };
 
-const includeTermCode = obj => (keyExists('term', obj) ? { ...obj, termCode: obj.term } : obj);
+const includeTermCode = (obj) => (keyExists('term', obj) ? { ...obj, termCode: obj.term } : obj);
 
-const makeHumanReadableTerm = obj => (keyExists('term', obj) ? { ...obj, term: getFullTermName(obj.term) } : obj);
+const makeHumanReadableTerm = (obj) =>
+  keyExists('term', obj) ? { ...obj, term: getFullTermName(obj.term) } : obj;
 
-const preferredNameAsFirstName = obj => (keyExists('firstName', obj) ? { ...obj, firstName: obj.preferredName || obj.firstName } : obj);
+const preferredNameAsFirstName = (obj) =>
+  keyExists('firstName', obj) ? { ...obj, firstName: obj.preferredName || obj.firstName } : obj;
 
-const addOpenDate = obj => (keyExists('startDate', obj)
-  ? {
-    ...obj,
-    openDate: getSundayBefore(obj.startDate),
-  }
-  : obj);
+const addOpenDate = (obj) =>
+  keyExists('startDate', obj)
+    ? {
+        ...obj,
+        openDate: getSundayBefore(obj.startDate),
+      }
+    : obj;
 
-const prettifyDateField = fieldName => obj => (keyExists(fieldName, obj)
-  ? {
-    ...obj,
-    [fieldName]: toPrettyDate(obj[fieldName]),
-  }
-  : obj);
+const prettifyDateField = (fieldName) => (obj) =>
+  keyExists(fieldName, obj)
+    ? {
+        ...obj,
+        [fieldName]: toPrettyDate(obj[fieldName]),
+      }
+    : obj;
 
 function cleanJexData(data) {
   if (Array.isArray(data)) return data.map(cleanJexData);
 
-  return R.pipe(
+  return pipe(
     toHumanReadableCredits,
     includeCourseId,
     includeTermCode,
@@ -59,4 +63,4 @@ function cleanJexData(data) {
   )(data);
 }
 
-module.exports = cleanJexData;
+export default cleanJexData;
