@@ -1,4 +1,4 @@
-const cleanJexData = require('./cleanJexData');
+import cleanJexData from './cleanJexData.js';
 
 describe('cleanJexData', () => {
   it('makes credits human readable', () => {
