@@ -1,6 +1,6 @@
-const removeExtraneousWhitespace = require('./removeExtraneousWhitespace');
+import removeExtraneousWhitespace from './removeExtraneousWhitespace';
 
-module.exports = function cleanMessage(message) {
+export default function cleanMessage(message) {
   return Object.keys(message).reduce(
     (acc, key) => ({
       ...acc,
@@ -9,4 +9,4 @@ module.exports = function cleanMessage(message) {
     }),
     {},
   );
-};
+}
