@@ -1,5 +1,5 @@
-const { DateTime } = require('luxon');
-const { default: ow } = require('ow');
+import { DateTime } from 'luxon';
+import ow from 'ow';
 
 const getTimestamp = () => DateTime.local().toString();
 const emailSummary = (emails) =>
@@ -46,7 +46,7 @@ const emailsReport = (emailMessages) => {
   return ['✉️  EMAILS', ...emailMessages.map(emailToString)].join('\n\n');
 };
 
-module.exports = function createTaskReport({ taskName, emails, errors, opts }) {
+export default function createTaskReport({ taskName, emails, errors, opts }) {
   ow(taskName, ow.string);
   ow(emails, ow.array);
   ow(errors, ow.array);
@@ -63,4 +63,4 @@ module.exports = function createTaskReport({ taskName, emails, errors, opts }) {
     .split('\n')
     .map((str) => str.trim())
     .join('\n');
-};
+}
