@@ -1,6 +1,10 @@
-const path = require('path');
-const Email = require('email-templates');
-const R = require('ramda');
+import { join, dirname } from 'path';
+import { fileURLToPath } from 'url';
+import Email from 'email-templates';
+import { partition } from 'ramda';
+
+// eslint-disable-next-line no-underscore-dangle
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 /**
  * generates messages given a collection of data
@@ -13,7 +17,7 @@ const R = require('ramda');
  *  emails, and a list of errors which occurred while generating.
  */
 async function generateEmails({ template, records, to, from, bcc, requiredFields = [] }) {
-  const templatePath = path.join(__dirname, `../tasks/${template}`);
+  const templatePath = join(__dirname, `../tasks/${template}`);
   const email = new Email({
     views: {
       options: {
@@ -36,7 +40,7 @@ async function generateEmails({ template, records, to, from, bcc, requiredFields
   // records. Incomplete records are those which may have a "null" or
   // "undefined" someplace within one of the required fields.
   const hasRequiredFields = (record) => requiredFields.every((key) => !!record[key]);
-  const [completeRecords, incompleteRecords] = R.partition(hasRequiredFields, records);
+  const [completeRecords, incompleteRecords] = partition(hasRequiredFields, records);
 
   // add any incomplete records to our list of errors
   const errors = incompleteRecords.map((record) =>
@@ -67,4 +71,4 @@ async function generateEmails({ template, records, to, from, bcc, requiredFields
   return { emails, errors };
 }
 
-module.exports = generateEmails;
+export default generateEmails;
