@@ -1,7 +1,4 @@
-const generateEmails = require('./generateEmails');
-// const jex = require('../services/jex');
-
-// jest.mock(jex);
+import generateEmails from './generateEmails.js';
 
 const validRecords = [
   {
