@@ -1,6 +1,7 @@
-const toShortYear = require('./toShortYear');
-const toHyphenatedCourseCode = require('./toHyphenatedCourseCode');
+import toShortYear from './toShortYear.js';
+import toHyphenatedCourseCode from './toHyphenatedCourseCode.js';
 
-const getCourseId = ({ courseCode, term, year }) => `${toHyphenatedCourseCode(courseCode)}-${term}${toShortYear(year)}`;
+const getCourseId = ({ courseCode, term, year }) =>
+  `${toHyphenatedCourseCode(courseCode)}-${term}${toShortYear(year)}`;
 
-module.exports = getCourseId;
+export default getCourseId;
