@@ -1,4 +1,4 @@
-const getCourseId = require('./getCourseId');
+import getCourseId from './getCourseId.js';
 
 describe('getCourseId', () => {
   it('gives a hyphenated course ID string given a course code, term, and year', () => {
