@@ -1,9 +1,9 @@
-const { promisify } = require('util');
-const fs = require('fs');
-const { join } = require('path');
+import { promisify } from 'util';
+import { readdir, stat } from 'fs';
+import { join } from 'path';
 
-const readdirAsync = promisify(fs.readdir);
-const statAsync = promisify(fs.stat);
+const readdirAsync = promisify(readdir);
+const statAsync = promisify(stat);
 
 async function getDirectories(path) {
   const dirContents = await readdirAsync(path);
@@ -14,4 +14,4 @@ async function getDirectories(path) {
   });
 }
 
-module.exports = getDirectories;
+export default getDirectories;
