@@ -1,4 +1,4 @@
-const getDirectories = require('./getDirectories');
+import getDirectories from './getDirectories.js';
 
 describe('getDirectories', () => {
   it('gets a list of directories in a given path', async () => {
