@@ -5,4 +5,4 @@ const getFullTermName = (term) => {
   return null;
 };
 
-module.exports = getFullTermName;
+export default getFullTermName;
