@@ -1,4 +1,4 @@
-const getFullTermName = require('./getFullTermName');
+import getFullTermName from './getFullTermName.js';
 
 describe('getFullTermName', () => {
   it('translates semesters into full name', () => {
