@@ -1,4 +1,4 @@
-const getSundayBefore = require('./getSundayBefore');
+import getSundayBefore from './getSundayBefore.js';
 
 describe('getSundayBefore', () => {
   it('gets sunday before a given date', () => {
