@@ -1,6 +1,6 @@
 let logContents = [];
-module.exports = function log(msg) {
+export default function log(msg) {
   console.log(msg);
   logContents = [...logContents, msg];
   return logContents;
-};
+}
