@@ -1,12 +1,12 @@
-const fs = require('fs');
+import { readFile } from 'fs';
 
 function readFilePromisified(path) {
   return new Promise((resolve, reject) => {
-    fs.readFile(path, 'utf8', (err, data) => {
+    readFile(path, 'utf8', (err, data) => {
       if (err) return reject(err);
       return resolve(data);
     });
   });
 }
 
-module.exports = readFilePromisified;
+export default readFilePromisified;
