@@ -1,3 +1,3 @@
-module.exports = function removeExtraneousWhitespace(str) {
+export default function removeExtraneousWhitespace(str) {
   return str.trim().replace(/\s+/g, ' ');
-};
+}
