@@ -1,9 +1,9 @@
-const postmark = require('postmark');
-const settings = require('../settings');
+import postmark from 'postmark';
+import settings from '../settings.js';
 
 const mailClient = new postmark.ServerClient(settings.postmark.apiKey);
 
-module.exports = ({ to, from, subject, html, text, bcc }) =>
+export default ({ to, from, subject, html, text, bcc }) =>
   mailClient.sendEmail({
     To: to,
     From: from,
