@@ -1,3 +1,3 @@
-module.exports = function toHyphenatedCourseCode(courseCode) {
+export default function toHyphenatedCourseCode(courseCode) {
   return courseCode.replace(/\s+/g, '-');
-};
+}
