@@ -1,4 +1,4 @@
-const toHyphenatedCourseCode = require('./toHyphenatedCourseCode');
+import toHyphenatedCourseCode from './toHyphenatedCourseCode.js';
 
 describe('toHyphenatedCourseCode', () => {
   it('hyphenates a space-separated course code', () => {
