@@ -8,8 +8,8 @@
  * getJexYear({term: 'SP', realYear: 2018})
  * // returns 2017 since Spring of 2018 is in AY2017-18
  */
-module.exports = function toJexYear({ term, realYear }) {
+export default function toJexYear({ term, realYear }) {
   if (!term || !realYear) throw Error('missing term or realYear');
   const yearInt = Number.parseInt(realYear, 10);
   return term === 'FA' ? yearInt : yearInt - 1;
-};
+}
