@@ -1,7 +1,7 @@
-const { DateTime } = require('luxon');
+import { DateTime } from 'luxon';
 
-module.exports = function toPrettyDate(date) {
+export default function toPrettyDate(date) {
   const dt = DateTime.fromJSDate(date);
   const prettyDate = dt.toLocaleString(DateTime.DATE_MED);
   return `${dt.weekdayShort}, ${prettyDate}`;
-};
+}
