@@ -1,4 +1,4 @@
-const toPrettyDate = require('./toPrettyDate');
+import toPrettyDate from './toPrettyDate.js';
 
 describe('toPrettyDate', () => {
   it('returns a pretty date', () => {
