@@ -2,6 +2,6 @@
  * last two digits of a given year
  * @param {Int|String} year
  */
-module.exports = function toShortYear(year) {
+export default function toShortYear(year) {
   return year.toString().slice(-2);
-};
+}
