@@ -1,4 +1,4 @@
-const toShortYear = require('./toShortYear');
+import toShortYear from './toShortYear.js';
 
 describe('toShortYear', () => {
   it('returns a 2-digit version of the year', () => {
