@@ -1,6 +1,6 @@
-const toJexYear = require('./toJexYear');
+import toJexYear from './toJexYear.js';
 
-module.exports = function withOnlyCoursesSql({ baseQuery, sectionTable = 'sch', courses = [] }) {
+export default function withOnlyCoursesSql({ baseQuery, sectionTable = 'sch', courses = [] }) {
   // restrict to Canvas Courses
   const restrictionsSQL = courses
     .map(({ year, term, sections }) => {
@@ -14,4 +14,4 @@ module.exports = function withOnlyCoursesSql({ baseQuery, sectionTable = 'sch', 
     .join('\n or ');
 
   return baseQuery.concat(`and (${restrictionsSQL})`);
-};
+}
