@@ -1,10 +1,6 @@
-const toJexYear = require('./toJexYear');
+import toJexYear from './toJexYear.js';
 
-module.exports = function withoutCanvasCoursesSql({
-  baseQuery,
-  sectionTable = 'sch',
-  courses = [],
-}) {
+export default function withoutCanvasCoursesSql({ baseQuery, sectionTable = 'sch', courses = [] }) {
   if (!courses.length) return baseQuery;
 
   const restrictionsSQL = courses
@@ -19,4 +15,4 @@ module.exports = function withoutCanvasCoursesSql({
     .join('\n or ');
 
   return baseQuery.concat(`and not (${restrictionsSQL})`);
-};
+}
