@@ -1,5 +1,5 @@
-const withoutCoursesSql = require('./withoutCoursesSql');
-const removeExtraneousWhitespace = require('./removeExtraneousWhitespace');
+import withoutCoursesSql from './withoutCoursesSql.js';
+import removeExtraneousWhitespace from './removeExtraneousWhitespace.js';
 
 const baseQuery = `
   select id, crs_cde, term, year, firstName, lastName
