@@ -1,5 +1,5 @@
-const write = require('write');
-const log = require('./log');
+import write from 'write';
+import log from './log.js';
 
 const writeLogFile = async (fileDest, contents) => {
   try {
@@ -11,4 +11,4 @@ const writeLogFile = async (fileDest, contents) => {
   }
 };
 
-module.exports = writeLogFile;
+export default writeLogFile;
