@@ -1,8 +1,12 @@
-const path = require('path');
-const jex = require('../../services/jex');
-const cleanJexData = require('../../lib/cleanJexData');
-const generateEmails = require('../../lib/generateEmails');
-const { salesforce } = require('../../settings');
+import { basename, dirname } from 'path';
+import { fileURLToPath } from 'url';
+import jex from '../../services/jex.js';
+import cleanJexData from '../../lib/cleanJexData.js';
+import generateEmails from '../../lib/generateEmails.js';
+import settings from '../../settings.js';
+
+// eslint-disable-next-line no-underscore-dangle
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 const createSQL = ({ today }) => {
   // use cast(getdate() as date) to get only the date
@@ -61,10 +65,13 @@ where ss.room_cde = 'OL'
 
 async function task({ today }) {
   const sql = createSQL({ today });
-  const records = await jex.query(sql).then(cleanJexData);
+  const records = await jex
+    .query(sql)
+    .then(cleanJexData)
+    .catch((err) => console.error(err));
 
   return generateEmails({
-    template: path.basename(__dirname),
+    template: basename(__dirname),
     records,
     to: ({ firstName, lastName, personalEmail, mcadEmail }) =>
       [
@@ -72,9 +79,9 @@ async function task({ today }) {
         personalEmail ? `${firstName} ${lastName} <${personalEmail}>` : '',
       ].join(', '),
     from: () => 'MCAD Online Learning <online@mcad.edu>',
-    bcc: () => [salesforce.email, 'MCAD Online Learning <online@mcad.edu>'].join(','),
+    bcc: () => [settings.salesforce.email, 'MCAD Online Learning <online@mcad.edu>'].join(','),
     requiredFields: ['username', 'mcadEmail'],
   });
 }
 
-module.exports = task;
+export default task;
